// src/app/api/models/combo-response.ts
/* tslint:disable */
/* eslint-disable */
/* Code generated by ng-openapi-gen DO NOT EDIT. */

import { BookResponse } from '../models/book-response';
export interface ComboResponse {
  books?: Array<BookResponse> | null;
  comboId?: number;
  createdDate?: string | null;
  description?: string | null;
  discountPrice?: number;
  image?: string | null;
  name?: string | null;
  totalPrice?: number;
}